+++ src/main/java/service/ServiceStudentGroup.java
@@ -1,6 +1,7 @@
 package service;
 
 import entity.*;
+import exception.EntityInUseException;
 import repository.StudentGroupRepository;
 import java.util.List;
 import java.util.Optional;
@@ -27,6 +28,11 @@ public class ServiceStudentGroup implements ServiceStudentGroupInterface{
     }
 
     public void delete(Long id) {
+        StudentGroup studentGroup = studentGroupRepository
+                .findById(id).orElseThrow(() -> new IllegalArgumentException("Student group with id " + id + " does not exist"));
+        if (!studentGroup.getStudents().isEmpty()){
+            throw new EntityInUseException("This university has students, number: " + studentGroup.getStudents().size());
+        }
         studentGroupRepository.deleteById(id);
     }
